refactor(metronome): extract `MetronomePermissionRationale` widget
